Network the composed things properly
Now they can all talk to each other, yay.
Main frustration is that the api accepts `db` as the database host but
vis won't do the same with `api`, so I've resorted to writing in
http://localhost.

Also reminded myself that the ports section open to the host machine,
expose is for networks among the containers.