Change layout to use ELK
elk has a bundle of algorithms. `disco` does cise, then packs the
disconnected subgraphs closer together. The package uses an older
version of elkjs, which has an issue where `g` is undefined. A later
revision fixes this but is not pulled in, hence the overrides.

overrode disco config to use stress layout. The animation is a little
slow, everything stays grouped in the top corner for a few seconds while
it thinks. I have a feeling that if I gave it all the nodes+edges, then
did a layout, it might go smoother.

Refs:
- https://github.com/cytoscape/cytoscape.js-elk