Replace backend with ent/go implementation
Not a full replacement, still some entities to add. This uses an ORM and
code gen thing called `ent`. The basics of what an object is and what's
exposed to the graphql api is still manual but it magics everything else
away. Mostly.

Most of the generated files have a warning at the top not to edit them.
Don't do that. Just play with /ent/schemas/* to update the objects.

pretense.graphql is my custom additions, and the *.resolvers.go need to
be modified to add custom resolvers (after generating below)

`go generate .` will regenerate most of the relevant files when
modifying schemas