Improve upserting
creating an object with relations may cause an error in the unique
constraint in another table, in this case topic_contains. The sdk
doesn't know how to handle those, so we upsert the main object, then
update relations after.

This approach means I might accumulate albums on a topic with nothing
underneath, but unlikely, given the amount of usage it gets